Move things around and delete old scripts and doc to prepare for merge with dcs-bios-hub
This commit adapts the directory structure for the merge with the new DCS-BIOS Hub.

- move JSON files to src/control-reference-json
- move Lua scripts to src/dcs-lua
- remove documentation and README files (those need a major rewrite anyway)
- remove .cmd scripts, .jsonp files and the control reference extension for Google Chrome (replaced by dcs-bios-hub)

Note that the tree this is based on has been run through 'git filter-branch' to remove
the 'socat' and 'Programs' folders. While this does break existing history, it means we do not need to keep
~17 MiB of unnecessary binaries in the history forever.
If necessary, changes that are based on the old history can be merged
by exporting them as a patch set and applying that on top of the new history.